Created main.ipynb
This is a version of the project which includes sort of an author commentary. I created this so that I can dump my ideas and other things which won't be included in the final report. I also wish to re-write most of the project as most of it was written using A.I.